Sprint Mission 4 기본 요구사항 설계 완료

+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileBinaryContentRepository.java
@@ -105,7 +105,9 @@ public class FileBinaryContentRepository implements BinaryContentRepository {
     public void deleteById(UUID userId) {
         Path path = resolvePath(userId);
         try {
-            Files.delete(path);
+            if (Files.exists(path)) {
+                Files.delete(path);
+            }
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
